# rest-api/v1/generated/java/src/main/java/it/tramways/projects/api/v1/model/Configurable.java
/*
 * Tramways Roadmaps API
 * Tramways Roadmaps API
 *
 * The version of the OpenAPI document: 1.0.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package it.tramways.projects.api.v1.model;

import java.util.Objects;
import java.util.Arrays;
import com.google.gson.TypeAdapter;
import com.google.gson.annotations.JsonAdapter;
import com.google.gson.annotations.SerializedName;
import com.google.gson.stream.JsonReader;
import com.google.gson.stream.JsonWriter;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import it.tramways.projects.api.v1.model.CrossingLink;
import it.tramways.projects.api.v1.model.Lane;
import it.tramways.projects.api.v1.model.Property;
import it.tramways.projects.api.v1.model.RelevantPoint;
import java.io.IOException;
import java.util.ArrayList;
import java.util.List;

/**
 * Configurable
 */
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2022-01-14T00:56:01.025+01:00[Europe/Berlin]")
public class Configurable {
  public static final String SERIALIZED_NAME_ID = "id";
  @SerializedName(SERIALIZED_NAME_ID)
  private String id;

  public static final String SERIALIZED_NAME_CATEGORY = "category";
  @SerializedName(SERIALIZED_NAME_CATEGORY)
  private String category;

  public static final String SERIALIZED_NAME_CONFIGURABLE_TYPE = "configurableType";
  @SerializedName(SERIALIZED_NAME_CONFIGURABLE_TYPE)
  protected String configurableType;

  public static final String SERIALIZED_NAME_PROPS = "props";
  @SerializedName(SERIALIZED_NAME_PROPS)
  private List<Property> props = null;

  public Configurable() {
    this.configurableType = this.getClass().getSimpleName();
  }

  public Configurable id(String id) {
    
    this.id = id;
    return this;
  }

   /**
   * Get id
   * @return id
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")

  public String getId() {
    return id;
  }


  public void setId(String id) {
    this.id = id;
  }


  public Configurable category(String category) {
    
    this.category = category;
    return this;
  }

   /**
   * Get category
   * @return category
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")

  public String getCategory() {
    return category;
  }


  public void setCategory(String category) {
    this.category = category;
  }


  public Configurable configurableType(String configurableType) {
    
    this.configurableType = configurableType;
    return this;
  }

   /**
   * Get configurableType
   * @return configurableType
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")

  public String getConfigurableType() {
    return configurableType;
  }


  public void setConfigurableType(String configurableType) {
    this.configurableType = configurableType;
  }


  public Configurable props(List<Property> props) {
    
    this.props = props;
    return this;
  }

  public Configurable addPropsItem(Property propsItem) {
    if (this.props == null) {
      this.props = new ArrayList<Property>();
    }
    this.props.add(propsItem);
    return this;
  }

   /**
   * Get props
   * @return props
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")

  public List<Property> getProps() {
    return props;
  }


  public void setProps(List<Property> props) {
    this.props = props;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Configurable configurable = (Configurable) o;
    return Objects.equals(this.id, configurable.id) &&
        Objects.equals(this.category, configurable.category) &&
        Objects.equals(this.configurableType, configurable.configurableType) &&
        Objects.equals(this.props, configurable.props);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, category, configurableType, props);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Configurable {\n");
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    category: ").append(toIndentedString(category)).append("\n");
    sb.append("    configurableType: ").append(toIndentedString(configurableType)).append("\n");
    sb.append("    props: ").append(toIndentedString(props)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

}