# rest-api/v1/generated/javascript/it/tramways/projects/api/v1/model/crossing-link.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tramways Roadmaps API
 * Tramways Roadmaps API
 *
 * The version of the OpenAPI document: 1.0.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { Configurable } from './configurable';
import { LaneAllOf } from './lane-all-of';
import { Property } from './property';

/**
 * @type CrossingLink
 * @export
 */
export type CrossingLink = Configurable & LaneAllOf;