# rest-api/v1/generated/javascript/it/tramways/projects/api/v1/model/project-description.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tramways Roadmaps API
 * Tramways Roadmaps API
 *
 * The version of the OpenAPI document: 1.0.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ProjectDescription
 */
export interface ProjectDescription {
    /**
     * 
     * @type {string}
     * @memberof ProjectDescription
     */
    uuid?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDescription
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDescription
     */
    owner?: string;
}