# rest-api/v1/generated/javascript/it/tramways/projects/api/v1/model/relevant-point.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tramways Roadmaps API
 * Tramways Roadmaps API
 *
 * The version of the OpenAPI document: 1.0.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { Configurable } from './configurable';
import { CrossingLink } from './crossing-link';
import { Property } from './property';
import { RelevantPointAllOf } from './relevant-point-all-of';

/**
 * @type RelevantPoint
 * @export
 */
export type RelevantPoint = Configurable & RelevantPointAllOf;