# rest-api/v1/generated/javascript/it/tramways/projects/api/v1/model/string-property.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tramways Roadmaps API
 * Tramways Roadmaps API
 *
 * The version of the OpenAPI document: 1.0.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { Property } from './property';
import { StringPropertyAllOf } from './string-property-all-of';

/**
 * @type StringProperty
 * @export
 */
export type StringProperty = Property & StringPropertyAllOf;